requirements.txt
alembic==1.15.1
beautifulsoup4==4.13.3
blinker==1.9.0
click==8.1.8
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==20.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.0.1
soupsieve==2.6
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
